colision

+++ src/utilz/HelpMethods.java
@@ -4,7 +4,7 @@ import main.Game;
 
 public class HelpMethods {
 
-    public static boolean canMoveHere(float x, float y, int width, int height,int[][] lvlData){
+    public static boolean canMoveHere(float x, float y, float width, float height,int[][] lvlData){
 
         if(!isSolid(x,y,lvlData)){
             if(!isSolid(x+width,y+height,lvlData)){
